api_llm/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
requests==2.32.4
bs4==0.0.2
python-multipart==0.0.20
SQLAlchemy==2.0.41
python-dotenv==1.1.1
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.28
unstructured==0.18.9
unstructured[pdf]==0.18.9
langchain-unstructured[local]==0.1.5
langchain-chroma==0.2.5